Implemetn Authorization using Gate

+++ app/Http/Controllers/QuestionsController.php
@@ -15,7 +15,7 @@ class QuestionsController extends Controller
     public function index()
     {
         //\DB::enableQueryLog();
-        $questions = Question::with('user')->latest()->paginate(10);
+        $questions = Question::with('user')->latest()->paginate(300);
         return view('questions.index')->with('questions', $questions);//->render();
         //  dd(\DB::getQueryLog());
     }
@@ -65,6 +65,10 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if(\Gate::denies('update-question', $question))
+        {
+            abort(403, 'Access denied');
+        }
         return view('questions.edit')->with('question', $question);
     }
 
@@ -77,6 +81,10 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question', $question))
+        {
+            abort(403, 'Access denied');
+        }
         $question->update($request->only('title', 'body'));
         return redirect()->route('questions.index')->with('success', 'The question has been updated successfully');
     }
@@ -89,6 +97,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question', $question))
+        {
+            abort(403, 'Access denied');
+        }
         $question->delete();
         return redirect()->route('questions.index')->with('success', 'question deleted successfully.');
     }
